* develop: (63 commits)
  Resync model getAttribute override with base Laravel functionality
  Convert Markdown parser to CommonMark (#133)
  Improve support for multiple database connections (#132)
  Make the SectionParser more extendable (#131)
  Fix static analysis errors
  Fixed generation of thumbnails for remote disks
  Get local root path from configured disk
  Use named arguments
  add more testing with pivot data
  Add Str::isJson() | is_json() helpers
  Code analysis fixes
  Register slug rule as part of Validation singleton registration
  Add slug validation rule
  Use Laravel's CLI components
  Use Laravel's CLI components
  Prioritize local dynamic methods over behavior-provided methods (#130)
  Delete unneeded PHPUnit config
  Re-enable code analysis on develop branch
  Fix PHPStan testing, minor tweaks to docs
  Pass the full model through add() and remove() methods
  ...

+++ src/Database/Traits/ArraySource.php
@@ -100,11 +100,13 @@ trait ArraySource
      */
     protected function arraySourceCreateDb(): void
     {
-        if (File::exists($this->arraySourceGetDbPath())) {
-            File::delete($this->arraySourceGetDbPath());
+        if ($this->arraySourceCanStoreDb()) {
+            if (File::exists($this->arraySourceGetDbPath())) {
+                File::delete($this->arraySourceGetDbPath());
+            }
+            // Create SQLite file
+            File::put($this->arraySourceGetDbPath(), '');
         }
-        // Create SQLite file
-        File::put($this->arraySourceGetDbPath(), '');
 
         $records = $this->getRecords();
